# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SU'UD Job Portal API",
        "description": "Complete API documentation for the SU'UD Job Portal platform. This API provides comprehensive job management, user authentication, application tracking, and administrative features for a modern job portal platform.",
        "contact": {
            "name": "SU'UD Development Team",
            "email": "admin@suud.sa"
        },
        "license": {
            "name": "Private License",
            "url": ""
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.suud.sa",
            "description": "Production Server"
        }
    ],
    "paths": {
        "/api/jobs": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Get all job listings",
                "description": "Retrieve a paginated list of active job listings with optional filtering",
                "operationId": "63963513839a812f56fe5eddf9fe529a",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term for job title or description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "description": "Filter by job location",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "job_type",
                        "in": "query",
                        "description": "Filter by job type",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "full_time",
                                "part_time",
                                "contract",
                                "internship"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (max 50)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jobs retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Create a new job listing",
                "description": "Create a new job listing (employers only)",
                "operationId": "c964c62023b02e106a37b155aeee84cc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "requirements",
                                    "location",
                                    "job_type"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Senior Software Developer"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "We are looking for a passionate developer..."
                                    },
                                    "requirements": {
                                        "type": "string",
                                        "example": "5+ years experience with React..."
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Riyadh, Saudi Arabia"
                                    },
                                    "job_type": {
                                        "type": "string",
                                        "enum": [
                                            "full-time",
                                            "part-time",
                                            "contract",
                                            "internship"
                                        ]
                                    },
                                    "experience_level": {
                                        "type": "string",
                                        "enum": [
                                            "entry",
                                            "mid",
                                            "senior",
                                            "executive"
                                        ]
                                    },
                                    "salary_min": {
                                        "type": "number",
                                        "example": 5000
                                    },
                                    "salary_max": {
                                        "type": "number",
                                        "example": 8000
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Technology"
                                    },
                                    "application_deadline": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "positions_available": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "skills": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "draft",
                                            "pending"
                                        ],
                                        "example": "pending"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Job created successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/jobs/{job:slug}": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Get job details",
                "description": "Get detailed information for a single job by slug",
                "operationId": "911b7c0044b9ab29c85e6c95a0f8d7dc",
                "parameters": [
                    {
                        "name": "job:slug",
                        "in": "path",
                        "description": "Job slug identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job not found or no longer available",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Job not found or no longer available"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/jobs/recent": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Get recent jobs",
                "description": "Get recent job listings for homepage display",
                "operationId": "a00922482ba9bcdeb1c1116b8ecb3173",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of jobs to return (max 10)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 10,
                            "example": 4
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recent jobs retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "jobs": {
                                                    "type": "array",
                                                    "items": {}
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/employer/jobs": {
            "get": {
                "tags": [
                    "Employer"
                ],
                "summary": "Get Employer's Job Listings",
                "description": "Get all job listings posted by the authenticated employer's company",
                "operationId": "db4d2f1ffc0e3170ea2986e01e689e71",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by job status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "active",
                                "declined",
                                "closed"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search in job title and description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (max 50)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Employer jobs retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "jobs": {
                                                    "type": "array",
                                                    "items": {}
                                                },
                                                "pagination": {
                                                    "properties": {
                                                        "current_page": {
                                                            "type": "integer"
                                                        },
                                                        "last_page": {
                                                            "type": "integer"
                                                        },
                                                        "per_page": {
                                                            "type": "integer"
                                                        },
                                                        "total": {
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied - Employer role required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Only employers can access job listings"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/admin/jobs/{job}/approve": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Approve Job Listing",
                "description": "Approve a pending job listing to make it active and visible to job seekers",
                "operationId": "304db953b159b76c9bef0b7b3cc1a7dd",
                "parameters": [
                    {
                        "name": "job",
                        "in": "path",
                        "description": "Job ID to approve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job approved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Job approved successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied - Admin role required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Only admins can approve job listings"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid job status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Only pending jobs can be approved"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/jobs/{job}/decline": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Decline Job Listing",
                "description": "Decline a pending job listing with a reason",
                "operationId": "80cdff0dd0021c462c31815d0a731ac6",
                "parameters": [
                    {
                        "name": "job",
                        "in": "path",
                        "description": "Job ID to decline",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reason"
                                ],
                                "properties": {
                                    "reason": {
                                        "type": "string",
                                        "maxLength": 500,
                                        "example": "Job requirements are not clear enough"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Job declined successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Job declined successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied - Admin role required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Only admins can decline job listings"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}